Adjust tab margin, underline position
